+++ sdk/trace/provider.go
@@ -180,6 +180,20 @@ func (p *TracerProvider) ApplyConfig(cfg Config) {
 	p.config.Store(&c)
 }
 
+// Shutdown shuts down the span processors in the order they were registered
+func (p *TracerProvider) Shutdown() {
+	spss, ok := p.spanProcessors.Load().(spanProcessorStates)
+	if !ok || len(spss) == 0 {
+		return
+	}
+
+	for _, sps := range spss {
+		sps.state.Do(func() {
+			sps.sp.Shutdown()
+		})
+	}
+}
+
 // WithSyncer registers the exporter with the TracerProvider using a
 // SimpleSpanProcessor.
 func WithSyncer(e export.SpanExporter) TracerProviderOption {